refactor(application): split local controller orchestration
